chat.gatewayからserver変数削除
